requirements.txt
nba_api>=1.4
pandas>=2.0
numpy>=1.26
xgboost>=2.0
scikit-learn>=1.4
streamlit>=1.32
plotly>=5.20
pyyaml>=6.0
joblib>=1.3
httpx>=0.27
pyarrow>=14.0
